File splitting and RustDoc addition
