+++ src-tauri/Cargo.toml
@@ -30,6 +30,7 @@ regex = "1.10"
 anyhow = "1.0"
 lazy_static = "1.4"
 sha2 = "0.10"
+url = "2.5"
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
 tauri-plugin-single-instance = "2"
